Break verification out of FormalSpec.
This is more flexible.
