+++ src/main/scala/chisel3/formal/FormalSpec.scala
@@ -27,58 +27,11 @@ import chisel3.formal.tags.Formal
  * ```
  */
 abstract class FormalSpec extends FlatSpec {
-  val sbyLineNumberPattern = """.*:\s*\d+(?:\.\d+)?-(\d+)(?:\.\d+)?.*""".r
-  val firrtlLineNumberPattern = """.*//\s*@\[(.*)]\s*""".r
   private var counter = 0
   def verify(dutGen: () => RawModule): Unit = {
     it should s"work in instance $counter" taggedAs(Formal) in {
-      val result = Driver(dutGen)
-      val rtlLines = result.rtl.linesIterator.toArray
-      // write out symbiyosys output stream
-      new PrintWriter(s"build/${result.moduleName}_sby_output.log") {
-        write(result.output)
-        close()
-      }
-      // write out symbiyosys error stream
-      new PrintWriter(s"build/${result.moduleName}_sby_error.log") {
-        write(result.error)
-        close()
-      }
-      for (line <- result.output.linesIterator) {
-        checkLine(result, rtlLines, line)
-      }
+      FormalVerify(dutGen)
     }
     counter += 1
-  }
-
-  def checkLine(result: VerificationResult, rtlLines: Array[String],
-                line: String): Unit = {
-    val errorEncountered = line.toLowerCase.contains("error")
-    val assertFailed = line.toLowerCase.contains("assert failed")
-    val coverFailed = line.toLowerCase.contains("unreached cover statement")
-    val message = if (coverFailed) {
-      "Failed to reach cover statement at"
-    } else if (assertFailed) {
-      "Failed to assert condition at"
-    } else {
-      "Error encountered at"
-    }
-    if (assertFailed || coverFailed || errorEncountered) {
-      line match {
-        case sbyLineNumberPattern(rtlLineNumber) => {
-          val rtlLine = rtlLines(rtlLineNumber.toInt-1)
-          rtlLine match {
-            case firrtlLineNumberPattern(scalaLine) => {
-              assert(false, s"$message $scalaLine")
-            }
-            case _ => {
-              assert(false, s"$message ${result.rtlFilename}" +
-                s":$rtlLineNumber (no Scala line ref found): $rtlLine")
-            }
-          }
-        }
-        case _ => assert(false, s"$message $line")
-      }
-    }
   }
 }
